Fixed problem with zips containing entries with compression method 8
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@963 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/common/formats/pdb/PdbStream.h
@@ -36,7 +36,7 @@ public:
 	size_t read(char *buffer, size_t maxSize);
 	virtual void close();
 
-	void seek(size_t offset);
+	void seek(int offset);
 	size_t offset() const;
 	size_t sizeOfOpened();
